+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.1
+
+- @EnableReflection:
+  - Added `reflectionClassName`, `reflectionExtensionName`.
+- @ReflectionBridge:
+  - Added `bridgeExtensionName`, `reflectionClassNames`, `reflectionExtensionNames`.
+- Builder: improved console output and verbose mode. 
+- Improved tests.
+
 ## 1.0.0
 
 - Support for Class reflection: